Add PropertyEqualityComparer class
This class allows one to provide instances of IEqualityComparer<T> using just a
projection on the type T (and potentially another IEqualityComparer<U> where type
U is the return type of the projection).  This allows class instances to be
compared based on a property of the class or some calculation on the instance of
the class.